+++ extensions/data-workspace/src/common/utils.ts
@@ -3,12 +3,7 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as azdata from 'azdata';
 import * as fs from 'fs';
-
-export function createHorizontalContainer(view: azdata.ModelView, items: azdata.Component[]): azdata.FlexContainer {
-	return view.modelBuilder.flexContainer().withItems(items, { CSSStyles: { 'margin-right': '5px', 'margin-bottom': '10px' } }).withLayout({ flexFlow: 'row' }).component();
-}
 
 export async function directoryExist(directoryPath: string): Promise<boolean> {
 	const stats = await getFileStatus(directoryPath);
